Get analytics data from redux
